Test fixes
There are a number of attempts to use Deci in ints, which won't work in
2.7 or 3.x, we fix that.

Description doesn't exist in symposion.  So that fails our tests pretty
hard.  Switch that to Private Abstract.  It's clear these tests have not
been run in a very long time.  So both failures and especially passes
need to be taken with salt.